+++ CHANGELOG.md
@@ -1,3 +1,88 @@
+<a name="0.11.2-ghost.4"></a>
+## <small>0.11.2-ghost.4 (2019-06-03)</small>
+
+* 🐛 Fixed range#expandByMarker not expanding to beginning/end of section (#677) ([0000d1d](https://github.com/bustle/mobiledoc-kit/commit/0000d1d)), closes [#677](https://github.com/bustle/mobiledoc-kit/issues/677)
+* Copy update in demo ([77de471](https://github.com/bustle/mobiledoc-kit/commit/77de471))
+* Fix sectionParser obliterating plain text content in certain circumstances (#685) ([e5f877f](https://github.com/bustle/mobiledoc-kit/commit/e5f877f)), closes [#685](https://github.com/bustle/mobiledoc-kit/issues/685)
+
+
+
+<a name="0.11.2-ghost.3"></a>
+## <small>0.11.2-ghost.3 (2019-05-04)</small>
+
+* 🐛 Fixed parser plugin handling of top-level comment nodes ([4696bfe](https://github.com/bustle/mobiledoc-kit/commit/4696bfe))
+* v0.11.2-ghost.3 ([228e2e4](https://github.com/bustle/mobiledoc-kit/commit/228e2e4))
+
+
+
+<a name="0.11.2-ghost.2"></a>
+## <small>0.11.2-ghost.2 (2019-05-03)</small>
+
+* 🐛 Fixed atoms with no text value being removed when parsing top-level markerables ([8d5b337](https://github.com/bustle/mobiledoc-kit/commit/8d5b337))
+* 🐛 Fixed atoms with no text value being removed when parsing top-level markerables ([4828dd6](https://github.com/bustle/mobiledoc-kit/commit/4828dd6))
+* Uncomment "FIXME" Google Docs nested uls test ([9e4a4c8](https://github.com/bustle/mobiledoc-kit/commit/9e4a4c8))
+* Uncomment "FIXME" Google Docs nested uls test (#678) ([358cd9d](https://github.com/bustle/mobiledoc-kit/commit/358cd9d)), closes [#678](https://github.com/bustle/mobiledoc-kit/issues/678)
+* v0.11.2-ghost.2 ([b808ae7](https://github.com/bustle/mobiledoc-kit/commit/b808ae7))
+
+
+
+<a name="0.11.2-ghost.1"></a>
+## <small>0.11.2-ghost.1 (2019-04-29)</small>
+
+* Fixed range#expandByMarker not expanding to beginning/end of section ([ab01ecb](https://github.com/bustle/mobiledoc-kit/commit/ab01ecb))
+* v0.11.2-ghost.1 ([2a24702](https://github.com/bustle/mobiledoc-kit/commit/2a24702))
+
+
+
+<a name="0.11.2"></a>
+## <small>0.11.2 (2019-03-31)</small>
+
+* Bump travis to Node 8 ([ac8b89f](https://github.com/bustle/mobiledoc-kit/commit/ac8b89f))
+* Bump yarn deps ([f85346a](https://github.com/bustle/mobiledoc-kit/commit/f85346a))
+* Drop Safari from the base testem ([f1486a6](https://github.com/bustle/mobiledoc-kit/commit/f1486a6))
+* fix error parsing list sections containing text nodes from indented HTML ([002f0d6](https://github.com/bustle/mobiledoc-kit/commit/002f0d6))
+* handle parsing of lists when wrapped with indented html ([ca8c6c5](https://github.com/bustle/mobiledoc-kit/commit/ca8c6c5))
+* trim leading/trailing whitespace from sections that can occur when parsing indented HTML ([333c293](https://github.com/bustle/mobiledoc-kit/commit/333c293))
+* v0.11.2 ([2104893](https://github.com/bustle/mobiledoc-kit/commit/2104893))
+
+
+
+<a name="0.11.1-ghost.7"></a>
+## <small>0.11.1-ghost.7 (2019-02-22)</small>
+
+* fix error parsing list sections containing text nodes from indented HTML ([5b5296e](https://github.com/bustle/mobiledoc-kit/commit/5b5296e))
+* handle parsing of lists when wrapped with indented html ([b7fe611](https://github.com/bustle/mobiledoc-kit/commit/b7fe611))
+* trim leading/trailing whitespace from sections that can occur when parsing indented HTML ([07b6739](https://github.com/bustle/mobiledoc-kit/commit/07b6739))
+* v0.11.1-ghost.7 ([6014f3f](https://github.com/bustle/mobiledoc-kit/commit/6014f3f))
+
+
+
+<a name="0.11.1-ghost.6"></a>
+## <small>0.11.1-ghost.6 (2019-02-13)</small>
+
+* add failing test for #648 ([29ff9e7](https://github.com/bustle/mobiledoc-kit/commit/29ff9e7)), closes [#648](https://github.com/bustle/mobiledoc-kit/issues/648)
+* add failing test for list-items being appended as top-level sections ([548fcbf](https://github.com/bustle/mobiledoc-kit/commit/548fcbf))
+* add failing test for nested lists ([40ddff3](https://github.com/bustle/mobiledoc-kit/commit/40ddff3))
+* add failing test for nested lists of different types ([c88d6b8](https://github.com/bustle/mobiledoc-kit/commit/c88d6b8))
+* add test for paragraph in list item parsing behaviour ([be9f27b](https://github.com/bustle/mobiledoc-kit/commit/be9f27b))
+* add tests for dom and section parser ignoring blank markup sections ([100c5fb](https://github.com/bustle/mobiledoc-kit/commit/100c5fb))
+* Added more failing tests for 648 ([2eed96e](https://github.com/bustle/mobiledoc-kit/commit/2eed96e))
+* do not group consecutive list sections of different types ([45e4d27](https://github.com/bustle/mobiledoc-kit/commit/45e4d27))
+* Export MOBILEDOC_VERSION & mobiledocRenderers from index ([acbb829](https://github.com/bustle/mobiledoc-kit/commit/acbb829))
+* failing test for #656 ([4bcbec6](https://github.com/bustle/mobiledoc-kit/commit/4bcbec6)), closes [#656](https://github.com/bustle/mobiledoc-kit/issues/656) [#656](https://github.com/bustle/mobiledoc-kit/issues/656)
+* failing test for section parser handling consecutive lists of varying types ([66d2120](https://github.com/bustle/mobiledoc-kit/commit/66d2120))
+* Fix format readme missing word ([27b4588](https://github.com/bustle/mobiledoc-kit/commit/27b4588))
+* fix grouping of nested lists into single top-level list ([a5353ff](https://github.com/bustle/mobiledoc-kit/commit/a5353ff))
+* fix handling of nested list sections of different types ([2019c94](https://github.com/bustle/mobiledoc-kit/commit/2019c94))
+* fix list-items being added as top-level sections when breaking out of lists ([821bf2e](https://github.com/bustle/mobiledoc-kit/commit/821bf2e))
+* fixed SectionParser handling of paragraph following a list ([892b66a](https://github.com/bustle/mobiledoc-kit/commit/892b66a))
+* fixed SectionParser handling of paragraph following single-marker markup section ([c6ad8cd](https://github.com/bustle/mobiledoc-kit/commit/c6ad8cd))
+* Further improvements to tests ([83356fa](https://github.com/bustle/mobiledoc-kit/commit/83356fa))
+* handle nested section-creating elements correctly in SectionParser ([dfcadc3](https://github.com/bustle/mobiledoc-kit/commit/dfcadc3))
+* v0.11.1-ghost.6 ([8acf611](https://github.com/bustle/mobiledoc-kit/commit/8acf611))
+
+
+
 <a name="0.11.1-ghost.5"></a>
 ## <small>0.11.1-ghost.5 (2018-11-21)</small>
 
@@ -5,6 +90,7 @@
 * failing test for #656 ([db7f5c8](https://github.com/bustle/mobiledoc-kit/commit/db7f5c8)), closes [#656](https://github.com/bustle/mobiledoc-kit/issues/656) [#656](https://github.com/bustle/mobiledoc-kit/issues/656)
 * handle nested section-creating elements correctly in SectionParser ([d235c52](https://github.com/bustle/mobiledoc-kit/commit/d235c52))
 * Revert "Fix issue #592" ([62da2b7](https://github.com/bustle/mobiledoc-kit/commit/62da2b7)), closes [#592](https://github.com/bustle/mobiledoc-kit/issues/592)
+* v0.11.1-ghost.5 ([f34fdf5](https://github.com/bustle/mobiledoc-kit/commit/f34fdf5))